Splitting Utility in multiple classes
