+++ Kyoo.Tests/Utility/UtilityTests.cs
@@ -13,12 +13,23 @@ namespace Kyoo.Tests
 			Expression<Func<Show, int>> member = x => x.ID;
 			Expression<Func<Show, object>> memberCast = x => x.ID;
 
-			Assert.True(Utility.IsPropertyExpression(null));
+			Assert.False(Utility.IsPropertyExpression(null));
 			Assert.True(Utility.IsPropertyExpression(member));
 			Assert.True(Utility.IsPropertyExpression(memberCast));
 
 			Expression<Func<Show, object>> call = x => x.GetID("test");
 			Assert.False(Utility.IsPropertyExpression(call));
+		}
+		
+		[Fact]
+		public void GetPropertyName_Test()
+		{
+			Expression<Func<Show, int>> member = x => x.ID;
+			Expression<Func<Show, object>> memberCast = x => x.ID;
+
+			Assert.Equal("ID", Utility.GetPropertyName(member));
+			Assert.Equal("ID", Utility.GetPropertyName(memberCast));
+			Assert.Throws<ArgumentException>(() => Utility.GetPropertyName(null));
 		}
 	}
 }
